Moved get_qapp to glue.utils.qt
