+++ glue/app/qt/preferences.py
@@ -116,7 +116,7 @@ class PreferencesDialog(QtWidgets.QDialog):
 
 if __name__ == "__main__":
 
-    from glue.external.qt import get_qapp
+    from glue.utils.qt import get_qapp
     app = get_qapp()
     widget = PreferencesDialog()
     widget.show()
