+++ glue/core/qt/data_collection_model.py
@@ -544,7 +544,7 @@ class LabeledDelegate(QtWidgets.QStyledItemDelegate):
 
 if __name__ == "__main__":
 
-    from glue.external.qt import get_qapp
+    from glue.utils.qt import get_qapp
     from qtpy import QtWidgets
     from glue.core import Data, DataCollection
