+++ glue/core/qt/data_combo_helper.py
@@ -323,7 +323,7 @@ class DataCollectionComboHelper(BaseDataComboHelper):
 
 if __name__ == "__main__":
 
-    from glue.external.qt import get_qapp
+    from glue.utils.qt import get_qapp
 
     app = get_qapp()
