+++ glue/dialogs/common/qt/component_selector.py
@@ -130,7 +130,7 @@ CUSTOM_QWIDGETS.append(ComponentSelector)
 def main():  # pragma: no cover
     import glue
     import numpy as np
-    from glue.external.qt import get_qapp
+    from glue.utils.qt import get_qapp
 
     d = glue.core.Data(label="hi")
     d2 = glue.core.Data(label="there")
