+++ glue/dialogs/custom_component/qt/widget.py
@@ -281,6 +281,6 @@ def main():
             print('\t%s' % c)
 
 if __name__ == "__main__":
-    from glue.external.qt import get_qapp
+    from glue.utils.qt import get_qapp
     app = get_qapp()
     main()
