+++ glue/qt/__init__.py
@@ -6,4 +6,4 @@ warnings.warn("The glue.qt subpackage is deprecated - see the v0.7 release "
               "announcement for more details", GlueDeprecationWarning)
 
 # For compatibility
-from glue.external.qt import get_qapp
+from glue.utils.qt import get_qapp
