+++ glue/utils/qt/__init__.py
@@ -10,3 +10,4 @@ from .mixins import *
 from .mime import *
 from .python_list_model import *
 from .threading import *
+from .app import *
