+++ glue/utils/qt/app.py
@@ -1,0 +1,19 @@
+from qtpy import QtCore, QtGui, QtWidgets, PYQT5
+
+qapp = None
+
+def get_qapp(icon_path=None):
+    global qapp
+    qapp = QtWidgets.QApplication.instance()
+    if qapp is None:
+        qapp = QtWidgets.QApplication([''])
+        qapp.setQuitOnLastWindowClosed(True)
+        if icon_path is not None:
+            qapp.setWindowIcon(QtGui.QIcon(icon_path))
+
+    # Make sure we use high resolution icons with PyQt5 for HDPI
+    # displays. TODO: check impact on non-HDPI displays.
+    if PYQT5:
+        qapp.setAttribute(QtCore.Qt.AA_UseHighDpiPixmaps)
+
+    return qapp
