+++ glue/utils/qt/colors.py
@@ -207,7 +207,7 @@ CUSTOM_QWIDGETS.append(QColormapCombo)
 
 if __name__ == "__main__":
 
-    from glue.external.qt import get_qapp
+    from glue.utils.qt import get_qapp
 
     app = get_qapp()
