+++ glue/utils/qt/threading.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import, division, print_function
 
-from ...external.qt import QtCore
+from qtpy import QtCore
 
 __all__ = ['Worker']
